+++ src/lib/translate/label-designer-translations.interface.ts
@@ -2,6 +2,7 @@ export interface LabelDesignerTranslationsInterface {
   'Label Designer'?: string;
   'File'?: string;
   'New'?: string;
+  'Clear'?: string;
   'Save'?: string;
   'Save to file'?: string;
   'Edit'?: string;
@@ -94,5 +95,6 @@ export interface LabelDesignerTranslationsInterface {
   'Add unique values from data'?: string;
   'Making QR Code smaller than 10mm is possible but not recommended'?: string;
   'Are you sure that you want to start a new empty label?'?: string;
+  'Are you sure that you want to clear the data form the label?'?: string;
   'Do you want to discard the local changes?'?: string;
 }