+++ docs/CHANGELOG.md
@@ -1,5 +1,18 @@
+<a name="3.1.5"></a>
+# 3.1.5 (2019-09-05)
+
+### Features
+* **doc:** added link to demo to README.nd
+
+<a name="3.1.4"></a>
+# 3.1.4 (2019-09-05)
+
+### Features
+* **editor:** added menu item to clear data only on the editor
+* **generator:** pre-fill free text fields with the field value
+
 <a name="3.1.3"></a>
-# 3.1.3 (2019-09-?)
+# 3.1.3 (2019-09-03)
 
 ### Features
 * **core:** added method to convert data according to schema to match available fields.